requirements.txt
bcrypt==4.0.1
connexion==2.5.1
Flask==2.2.2
Flask-RESTful==0.3.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
Werkzeug==2.2.2
swagger-ui-bundle==0.0.9
flake8==6.0.0
flask-sqlalchemy==3.0.3
flask-migrate
psycopg2==2.9.5
uuid==1.30
python-dotenv==0.21.1